PYI-654: Update structure of the PassportCredentialIssuerResponse to store the various name fields within a 'names' property
